Skipping Bioconductor tests for R-devel (#28)
According to https://bioconductor.org/install/, the development version of Bioconductor (currently 3.14) works with R version 4.1.0. Since the R-devel check uses R 4.2, even tests using BiocManager devel will not work with R-devel. This is expected to be resolved soon, but I'd rather not count or depend on it and instead just skip Bioconductor badges on R-devel.